+++ js/preset-const.js
@@ -1,0 +1,60 @@
+const TITLES = [
+  'Апартаменты-студия, 19 м²',
+  'Квартира, 111 м²',
+  'Квартира, 350 м²'
+];
+
+const TYPES = [
+  'palace',
+  'flat',
+  'house',
+  'bungalow',
+  'hotel'
+];
+
+const CHECKIN = [
+  '12:00',
+  '13:00',
+  '14:00'
+];
+
+const CHECKOUT = [
+  '12:00',
+  '13:00',
+  '14:00'
+];
+
+const FEATURES = [
+  'wifi',
+  'dishwasher',
+  'parking',
+  'washer',
+  'elevator',
+  'conditioner'
+];
+
+const DESCRIPTIONS = [
+  'Рядом большое количество магазинов',
+  'До метро 1 мин пешком, или 2 остановки на трамвае, остановка у дома',
+  'Проблем с парковкой нет, идеальная транспортная доступность, до 3го кольца 10 мин на машине'
+];
+
+const PHOTOS = [
+  'https://assets.htmlacademy.ru/content/intensive/javascript-1/keksobooking/duonguyen-8LrGtIxxa4w.jpg',
+  'https://assets.htmlacademy.ru/content/intensive/javascript-1/keksobooking/brandon-hoogenboom-SNxQGWxZQi0.jpg',
+  'https://assets.htmlacademy.ru/content/intensive/javascript-1/keksobooking/claire-rendall-b6kAwr1i0Iw.jpg'
+];
+
+const SIMILAR_ADD_COUNT = 10;
+const MAX_PRICE = 1e9;
+const MAX_ROOMS_COUNT = 50;
+const MAX_GUESTS_NUMBER = 30;
+const MIN_LAT_POINT = 35.65000;
+const MAX_LAT_POINT = 35.70000;
+const MIN_LNG_POINT = 139.70000;
+const MAX_LNG_POINT = 139.80000;
+const COORDINATE_FRACTIONAL_DIGITS = 5;
+const AVATAR_PATH = 'img/avatars/user';
+
+export {TITLES, TYPES, CHECKIN, CHECKOUT, FEATURES, DESCRIPTIONS, PHOTOS, SIMILAR_ADD_COUNT, MAX_PRICE, MAX_ROOMS_COUNT, MAX_GUESTS_NUMBER, MIN_LAT_POINT,
+  MAX_LAT_POINT, MIN_LNG_POINT, MAX_LNG_POINT, COORDINATE_FRACTIONAL_DIGITS, AVATAR_PATH};
